+++ generators/gen_models.py
@@ -2,7 +2,7 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from s2gos_gen.common import (
+from generators.common import (
     S2GOS_PATH,
     OPEN_API_PATH,
 )