+++ generators/gen_server.py
@@ -6,7 +6,7 @@
 from pathlib import Path
 from typing import Literal
 
-from s2gos_gen.common import (
+from generators.common import (
     C_TAB,
     OPEN_API_PATH,
     S2GOS_PATH,
@@ -15,7 +15,7 @@ from s2gos_gen.common import (
     parse_responses,
     write_file,
 )
-from s2gos_gen.openapi import load_openapi_schema, OASchema, OAMethod
+from generators.openapi import load_openapi_schema, OASchema, OAMethod
 
 
 GENERATOR_NAME = str(Path(__file__).name)