+++ CHANGES.md
@@ -1,10 +1,9 @@
 ## Changes in version 0.0.2 (in development)
 
 * Reorganized package
-* Started code generation for models and client API
+* Started code generation for models, client, and server
+* Added basic FastAPI server, added server CLI `s2gos-server`
 
 ## Version 0.0.1 (not released) 
 
 Initial setup.
-
-